+++ bbtool/incremental/__init__.py
@@ -1,5 +1,6 @@
 from .cache import IncrementalCache
 from .manifest import (
+    campaign_identity_payload,
     find_previous_manifest,
     prune_manifests,
     write_manifest,
@@ -8,6 +9,7 @@ from .verify import first_difference
 
 __all__ = (
     "IncrementalCache",
+    "campaign_identity_payload",
     "find_previous_manifest",
     "first_difference",
     "prune_manifests",